Refactor caching store freshness and errors
