+++ internal/beads/caching_store.go
@@ -3,10 +3,7 @@ package beads
 import (
 	"context"
 	"encoding/json"
-	"errors"
 	"fmt"
-	"log"
-	"slices"
 	"sync"
 	"sync/atomic"
 	"time"
@@ -17,25 +14,20 @@ import (
 // through to the backing store and update the cache on success.
 //
 // External writes (agents running bd directly) are picked up via the
-// bd hook → gc event emit → event bus path. Call ApplyEvent when the
-// event bus delivers bead.created/updated/closed events. A background
-// reconciler periodically syncs as a safety net.
+// bd hook -> gc event emit -> event bus path. Call ApplyEvent when the
+// event bus delivers bead.created/updated/closed events. The background
+// reconciler acts as a watchdog and only performs a full scan once the
+// cache has gone stale or degraded.
 //
 // Only wraps BdStore because the event hook path requires dolt/bd.
 type CachingStore struct {
 	backing Store // runtime: always *BdStore; tests may use MemStore
 
-	mu       sync.RWMutex
-	beads    map[string]Bead
-	deps     map[string][]Dep
-	state    cacheState
-	syncedAt time.Time
-
-	// primeReady is closed when Prime completes (success or failure).
-	// Full-scan reads (List) block on this instead of falling through
-	// to a bd subprocess, preventing stampedes during startup.
-	primeReady chan struct{}
-	primeErr   error // set if Prime fails; readers fall through to backing
+	mu          sync.RWMutex
+	beads       map[string]Bead
+	deps        map[string][]Dep
+	state       cacheState
+	lastFreshAt time.Time
 
 	reconciling  atomic.Bool
 	syncFailures int
@@ -48,26 +40,31 @@ type cacheState int
 
 const (
 	cacheUninitialized cacheState = iota
-	cachePartial                  // PrimeActive loaded active beads; filtered active queries hit cache immediately
+	cachePartial                  // PrimeActive loaded active beads; active queries can use the cache immediately.
 	cacheLive
 	cacheDegraded
 )
 
-// CacheStats exposes reconciliation metrics for observability.
+// CacheStats exposes cache freshness, reconciliation, and problem state.
 type CacheStats struct {
 	TotalBeads      int
 	TotalDeps       int
+	LastFreshAt     time.Time
 	LastReconcileAt time.Time
 	LastReconcileMs float64
 	Adds            int64
 	Removes         int64
 	Updates         int64
 	SyncFailures    int
+	ProblemCount    int64
+	LastProblemAt   time.Time
+	LastProblem     string
 	State           string
 }
 
 const (
 	maxCacheSyncFailures         = 5
+	cacheReconcilePollInterval   = 5 * time.Second
 	cacheReconcileIntervalSmall  = 30 * time.Second
 	cacheReconcileIntervalMedium = 60 * time.Second
 	cacheReconcileIntervalLarge  = 120 * time.Second
@@ -75,11 +72,11 @@ const (
 
 // NewCachingStore wraps a BdStore with an in-memory read cache.
 // Call Prime() before serving reads, then StartReconciler() for
-// background sync. The onChange callback (optional) is called for
+// watchdog reconciliation. The onChange callback (optional) is called for
 // each detected external change with event type and bead JSON.
 //
-// Only BdStore is supported because the event hook path (bd hooks →
-// gc event emit → event bus → ApplyEvent) requires dolt infrastructure.
+// Only BdStore is supported because the event hook path (bd hooks ->
+// gc event emit -> event bus -> ApplyEvent) requires dolt infrastructure.
 func NewCachingStore(backing *BdStore, onChange func(eventType, beadID string, payload json.RawMessage)) *CachingStore {
 	return newCachingStore(backing, onChange)
 }
@@ -92,19 +89,17 @@ func NewCachingStoreForTest(backing Store, onChange func(eventType, beadID strin
 
 func newCachingStore(backing Store, onChange func(eventType, beadID string, payload json.RawMessage)) *CachingStore {
 	return &CachingStore{
-		backing:    backing,
-		beads:      make(map[string]Bead),
-		deps:       make(map[string][]Dep),
-		onChange:   onChange,
-		primeReady: make(chan struct{}),
+		backing:  backing,
+		beads:    make(map[string]Bead),
+		deps:     make(map[string][]Dep),
+		onChange: onChange,
 	}
 }
 
 // PrimeActive loads all non-closed beads (open + in_progress) into the
-// cache. These are fast indexed queries (~1-2s total) that populate
-// enough data for the startup path (adoption, session snapshot, desired
-// state) without waiting for the full Prime. The cache enters
-// cachePartial state: filtered List queries and Get hit cache for primed
+// cache. These are fast indexed queries that populate enough data for
+// startup paths without waiting for a full scan. The cache enters
+// cachePartial state: filtered active queries and Get hit cache for primed
 // beads, while closed-bead queries still delegate to the backing store.
 func (c *CachingStore) PrimeActive() error {
 	var all []Bead
@@ -124,12 +119,12 @@ func (c *CachingStore) PrimeActive() error {
 	if c.state == cacheUninitialized {
 		c.state = cachePartial
 	}
-	log.Printf("caching-store: pre-primed %d active beads", len(all))
+	c.markFreshLocked(time.Now())
+	c.updateStatsLocked()
 	return nil
 }
 
 // Prime loads all active beads and deps from the backing store into memory.
-// Closes the primeReady channel on completion so blocked callers can proceed.
 // Retries up to 3 times on failure since bd list can time out under
 // concurrent dolt load.
 func (c *CachingStore) Prime(_ context.Context) error {
@@ -140,14 +135,12 @@ func (c *CachingStore) Prime(_ context.Context) error {
 		if err == nil {
 			break
 		}
-		log.Printf("caching-store: prime attempt %d/3 failed: %v", attempt, err)
+		c.recordProblem(fmt.Sprintf("prime cache attempt %d/3", attempt), err)
 		if attempt < 3 {
 			time.Sleep(time.Duration(attempt*5) * time.Second)
 		}
 	}
 	if err != nil {
-		c.primeErr = err
-		c.signalPrimeReady()
 		return fmt.Errorf("prime list: %w", err)
 	}
 
@@ -155,44 +148,26 @@ func (c *CachingStore) Prime(_ context.Context) error {
 	for _, b := range all {
 		beadMap[b.ID] = cloneBead(b)
 	}
-	// Batch-fetch deps in one subprocess call (if backing is BdStore).
-	depMap := make(map[string][]Dep)
-	if bdStore, ok := c.backing.(*BdStore); ok && len(beadMap) > 0 {
-		ids := make([]string, 0, len(beadMap))
-		for id := range beadMap {
-			ids = append(ids, id)
-		}
-		if batchDeps, err := bdStore.DepListBatch(ids); err == nil {
-			for id, deps := range batchDeps {
-				depMap[id] = slices.Clone(deps)
-			}
-		}
+
+	depMap, depErr := c.fetchDepsForIDs(beadIDs(beadMap))
+	if depErr != nil {
+		c.recordProblem("prime dep cache", depErr)
 	}
 
+	now := time.Now()
 	c.mu.Lock()
 	defer c.mu.Unlock()
 	c.beads = beadMap
 	c.deps = depMap
 	c.state = cacheLive
-	c.syncedAt = time.Now()
 	c.syncFailures = 0
+	c.stats.SyncFailures = 0
+	c.markFreshLocked(now)
 	c.updateStatsLocked()
-	log.Printf("caching-store: primed %d beads, %d dep entries", len(beadMap), len(depMap))
-	c.signalPrimeReady()
 	return nil
 }
 
-// signalPrimeReady closes the primeReady channel exactly once.
-func (c *CachingStore) signalPrimeReady() {
-	select {
-	case <-c.primeReady:
-		// Already closed (from a previous Prime call).
-	default:
-		close(c.primeReady)
-	}
-}
-
-// StartReconciler launches background periodic sync. Cancel ctx to stop.
+// StartReconciler launches watchdog reconciliation. Cancel ctx to stop.
 func (c *CachingStore) StartReconciler(ctx context.Context) {
 	ctx, cancel := context.WithCancel(ctx)
 	c.cancelFn = cancel
@@ -210,8 +185,11 @@ func (c *CachingStore) StopReconciler() {
 func (c *CachingStore) Stats() CacheStats {
 	c.mu.RLock()
 	defer c.mu.RUnlock()
+
 	s := c.stats
 	switch c.state {
+	case cachePartial:
+		s.State = "partial"
 	case cacheLive:
 		s.State = "live"
 	case cacheDegraded:
@@ -232,604 +210,27 @@ func (c *CachingStore) IsLive() bool {
 // Backing returns the underlying store.
 func (c *CachingStore) Backing() Store { return c.backing }
 
-// ApplyEvent updates the cache from a bd hook event. Call this when the
-// event bus delivers a bead.created, bead.updated, or bead.closed event
-// with the full bead JSON payload. This provides sub-second cache
-// freshness for agent-initiated bd mutations without waiting for
-// reconciliation.
-func (c *CachingStore) ApplyEvent(eventType string, payload json.RawMessage) {
-	if len(payload) == 0 {
-		return
-	}
-	var b Bead
-	if err := json.Unmarshal(payload, &b); err != nil {
-		return
-	}
-	if b.ID == "" {
-		return
-	}
-	// bd hook payloads use "issue_type" while Bead uses "type". If Type
-	// wasn't populated from "type", check for the bd field name.
-	if b.Type == "" {
-		var bdCompat struct {
-			IssueType string `json:"issue_type"`
-		}
-		if err := json.Unmarshal(payload, &bdCompat); err == nil && bdCompat.IssueType != "" {
-			b.Type = bdCompat.IssueType
-		}
-	}
+func (c *CachingStore) markFreshLocked(now time.Time) {
+	c.lastFreshAt = now
+	c.stats.LastFreshAt = now
+}
 
+func (c *CachingStore) recordProblem(op string, err error) {
+	if err == nil {
+		return
+	}
 	c.mu.Lock()
 	defer c.mu.Unlock()
-	if c.state != cacheLive {
+	c.recordProblemLocked(op, err)
+}
+
+func (c *CachingStore) recordProblemLocked(op string, err error) {
+	if err == nil {
 		return
 	}
-
-	switch eventType {
-	case "bead.created":
-		if _, exists := c.beads[b.ID]; !exists {
-			c.beads[b.ID] = cloneBead(b)
-			c.updateStatsLocked()
-		}
-	case "bead.updated":
-		c.beads[b.ID] = cloneBead(b)
-	case "bead.closed":
-		if _, exists := c.beads[b.ID]; !exists {
-			c.updateStatsLocked()
-		}
-		c.beads[b.ID] = cloneBead(b)
-	}
-}
-
-// ApplyDepEvent updates the dep cache for a bead. Call after dep
-// mutations are detected via events or write-through.
-func (c *CachingStore) ApplyDepEvent(beadID string, deps []Dep) {
-	c.mu.Lock()
-	defer c.mu.Unlock()
-	if c.state != cacheLive {
-		return
-	}
-	c.deps[beadID] = slices.Clone(deps)
-}
-
-// ── Read methods (cache when live, fallback to backing) ─────────────
-
-// List returns beads matching the query. Active-bead queries are served from
-// cache when available. IncludeClosed queries merge cached active results with
-// backing-store history when possible so callers keep the old best-effort
-// behavior from ListByLabel/ListByMetadata during transient bd failures.
-func (c *CachingStore) List(query ListQuery) ([]Bead, error) {
-	if !query.HasFilter() && !query.AllowScan {
-		return nil, fmt.Errorf("listing beads: %w", ErrQueryRequiresScan)
-	}
-
-	c.mu.RLock()
-	state := c.state
-	if state == cacheLive || state == cachePartial {
-		// PrimeActive loads the full active set (open + in_progress), so
-		// active-only queries are complete even before the history prime finishes.
-		cached := make([]Bead, 0, len(c.beads))
-		for _, b := range c.beads {
-			if !query.Matches(b) {
-				continue
-			}
-			cached = append(cached, cloneBead(b))
-		}
-		c.mu.RUnlock()
-
-		finish := func(items []Bead) ([]Bead, error) {
-			sortBeadsForQuery(items, query.Sort)
-			if query.Limit > 0 && len(items) > query.Limit {
-				items = items[:query.Limit]
-			}
-			return items, nil
-		}
-
-		if !query.IncludesClosed() {
-			return finish(cached)
-		}
-
-		// The cache never has a complete closed-only or parent-history view, so
-		// preserve the old backing-store behavior for those query shapes.
-		if query.Status == "closed" || query.ParentID != "" {
-			return c.backing.List(query)
-		}
-
-		all, err := c.backing.List(query)
-		if err != nil {
-			return finish(cached)
-		}
-
-		seen := make(map[string]bool, len(cached))
-		for _, b := range cached {
-			seen[b.ID] = true
-		}
-		for _, b := range all {
-			if seen[b.ID] {
-				continue
-			}
-			cached = append(cached, b)
-			seen[b.ID] = true
-		}
-		return finish(cached)
-	}
-	c.mu.RUnlock()
-	return c.backing.List(query)
-}
-
-// ListOpen returns all cached beads, optionally filtered by status.
-func (c *CachingStore) ListOpen(status ...string) ([]Bead, error) {
-	query := ListQuery{AllowScan: true}
-	if len(status) > 0 {
-		query.Status = status[0]
-	}
-	return c.List(query)
-}
-
-// Get returns a single bead by ID from the cache or backing store.
-func (c *CachingStore) Get(id string) (Bead, error) {
-	c.mu.RLock()
-	if c.state == cacheLive || c.state == cachePartial {
-		if b, ok := c.beads[id]; ok {
-			c.mu.RUnlock()
-			return cloneBead(b), nil
-		}
-		c.mu.RUnlock()
-		return c.backing.Get(id)
-	}
-	c.mu.RUnlock()
-	return c.backing.Get(id)
-}
-
-// Ready returns open beads whose blocking deps are all closed.
-func (c *CachingStore) Ready() ([]Bead, error) {
-	c.mu.RLock()
-	if c.state == cacheLive {
-		statusByID := make(map[string]string, len(c.beads))
-		depsByID := make(map[string][]Dep, len(c.deps))
-		openBeads := make([]Bead, 0, len(c.beads))
-		missingDepIDs := make(map[string]struct{})
-		for _, b := range c.beads {
-			statusByID[b.ID] = b.Status
-			if b.Status == "open" && !IsReadyExcludedType(b.Type) {
-				openBeads = append(openBeads, cloneBead(b))
-			}
-		}
-		for _, b := range openBeads {
-			deps := slices.Clone(c.deps[b.ID])
-			depsByID[b.ID] = deps
-			for _, dep := range deps {
-				switch dep.Type {
-				case "blocks", "waits-for", "conditional-blocks":
-				default:
-					continue
-				}
-				if _, ok := statusByID[dep.DependsOnID]; !ok {
-					missingDepIDs[dep.DependsOnID] = struct{}{}
-				}
-			}
-		}
-		c.mu.RUnlock()
-
-		for depID := range missingDepIDs {
-			dep, err := c.backing.Get(depID)
-			if err != nil {
-				if errors.Is(err, ErrNotFound) {
-					continue
-				}
-				return nil, err
-			}
-			statusByID[depID] = dep.Status
-		}
-
-		var result []Bead
-		for _, b := range openBeads {
-			blocked := false
-			for _, dep := range depsByID[b.ID] {
-				switch dep.Type {
-				case "blocks", "waits-for", "conditional-blocks":
-				default:
-					continue
-				}
-				if statusByID[dep.DependsOnID] != "closed" {
-					blocked = true
-					break
-				}
-			}
-			if !blocked {
-				result = append(result, cloneBead(b))
-			}
-		}
-		return result, nil
-	}
-	c.mu.RUnlock()
-	return c.backing.Ready()
-}
-
-// Children returns beads with the given parent ID.
-func (c *CachingStore) Children(parentID string, opts ...QueryOpt) ([]Bead, error) {
-	return c.List(ListQuery{
-		ParentID:      parentID,
-		IncludeClosed: HasOpt(opts, IncludeClosed),
-		Sort:          SortCreatedAsc,
-	})
-}
-
-// ListByLabel returns beads matching the given label. By default, serves from
-// cache only (non-closed beads). Pass IncludeClosed to also query the backing
-// store for closed beads and merge results.
-func (c *CachingStore) ListByLabel(label string, limit int, opts ...QueryOpt) ([]Bead, error) {
-	return c.List(ListQuery{
-		Label:         label,
-		Limit:         limit,
-		IncludeClosed: HasOpt(opts, IncludeClosed),
-		Sort:          SortCreatedDesc,
-	})
-}
-
-// ListByAssignee returns beads assigned to the given agent with matching status.
-func (c *CachingStore) ListByAssignee(assignee, status string, limit int) ([]Bead, error) {
-	return c.List(ListQuery{
-		Assignee: assignee,
-		Status:   status,
-		Limit:    limit,
-		Sort:     SortCreatedDesc,
-	})
-}
-
-// ListByMetadata filters beads by metadata key-value pairs. By default, serves
-// from cache only (non-closed beads). Pass IncludeClosed to also query the
-// backing store for closed beads and merge results.
-func (c *CachingStore) ListByMetadata(filters map[string]string, limit int, opts ...QueryOpt) ([]Bead, error) {
-	return c.List(ListQuery{
-		Metadata:      filters,
-		Limit:         limit,
-		IncludeClosed: HasOpt(opts, IncludeClosed),
-		Sort:          SortCreatedDesc,
-	})
-}
-
-func matchesMetadata(b Bead, filters map[string]string) bool {
-	for k, v := range filters {
-		if b.Metadata[k] != v {
-			return false
-		}
-	}
-	return true
-}
-
-// DepList returns dependencies for a bead in the given direction.
-func (c *CachingStore) DepList(id, direction string) ([]Dep, error) {
-	c.mu.RLock()
-	if c.state == cacheLive {
-		if direction == "down" || direction == "" {
-			if deps, ok := c.deps[id]; ok {
-				c.mu.RUnlock()
-				return slices.Clone(deps), nil
-			}
-			// Dep not cached yet — fetch from backing and cache it.
-			c.mu.RUnlock()
-			deps, err := c.backing.DepList(id, direction)
-			if err != nil {
-				return nil, err
-			}
-			c.mu.Lock()
-			c.deps[id] = slices.Clone(deps)
-			c.mu.Unlock()
-			return deps, nil
-		}
-		// Reverse lookups are only partially cached; defer to the backing
-		// store so callers do not observe incomplete results.
-		c.mu.RUnlock()
-		return c.backing.DepList(id, direction)
-	}
-	c.mu.RUnlock()
-	return c.backing.DepList(id, direction)
-}
-
-// Ping delegates to the backing store.
-func (c *CachingStore) Ping() error {
-	return c.backing.Ping()
-}
-
-// ── Write methods (pass through + update cache) ─────────────────────
-
-// Create passes through to the backing store and updates the cache.
-func (c *CachingStore) Create(b Bead) (Bead, error) {
-	created, err := c.backing.Create(b)
-	if err != nil {
-		return created, err
-	}
-	c.mu.Lock()
-	c.beads[created.ID] = cloneBead(created)
-	c.mu.Unlock()
-	c.notifyChange("bead.created", created)
-	return created, nil
-}
-
-// Update passes through to the backing store and refreshes the cache.
-func (c *CachingStore) Update(id string, opts UpdateOpts) error {
-	if err := c.backing.Update(id, opts); err != nil {
-		return err
-	}
-	// Re-fetch from backing to get the authoritative state.
-	if fresh, err := c.backing.Get(id); err == nil {
-		c.mu.Lock()
-		c.beads[id] = cloneBead(fresh)
-		c.mu.Unlock()
-		c.notifyChange("bead.updated", fresh)
-	} else {
-		c.mu.Lock()
-		delete(c.beads, id)
-		delete(c.deps, id)
-		c.mu.Unlock()
-		log.Printf("caching-store: refresh after update failed for %s: %v", id, err)
-	}
-	return nil
-}
-
-// Close marks a bead as closed in the backing store and cache.
-func (c *CachingStore) Close(id string) error {
-	if err := c.backing.Close(id); err != nil {
-		return err
-	}
-	c.mu.Lock()
-	if b, ok := c.beads[id]; ok {
-		b.Status = "closed"
-		c.beads[id] = b
-		c.mu.Unlock()
-		c.notifyChange("bead.closed", b)
-	} else {
-		c.mu.Unlock()
-	}
-	return nil
-}
-
-// CloseAll closes multiple beads and sets metadata on each.
-func (c *CachingStore) CloseAll(ids []string, metadata map[string]string) (int, error) {
-	n, err := c.backing.CloseAll(ids, metadata)
-	if err != nil {
-		return n, err
-	}
-	c.mu.Lock()
-	for _, id := range ids {
-		if b, ok := c.beads[id]; ok {
-			b.Status = "closed"
-			if b.Metadata == nil {
-				b.Metadata = make(map[string]string, len(metadata))
-			}
-			for k, v := range metadata {
-				b.Metadata[k] = v
-			}
-			c.beads[id] = b
-		}
-	}
-	c.mu.Unlock()
-	return n, nil
-}
-
-// SetMetadata sets a single metadata key-value on a bead.
-func (c *CachingStore) SetMetadata(id, key, value string) error {
-	if err := c.backing.SetMetadata(id, key, value); err != nil {
-		return err
-	}
-	c.mu.Lock()
-	if b, ok := c.beads[id]; ok {
-		if b.Metadata == nil {
-			b.Metadata = make(map[string]string)
-		}
-		b.Metadata[key] = value
-		c.beads[id] = b
-	}
-	c.mu.Unlock()
-	return nil
-}
-
-// SetMetadataBatch sets multiple metadata key-values on a bead.
-func (c *CachingStore) SetMetadataBatch(id string, kvs map[string]string) error {
-	if err := c.backing.SetMetadataBatch(id, kvs); err != nil {
-		return err
-	}
-	c.mu.Lock()
-	if b, ok := c.beads[id]; ok {
-		if b.Metadata == nil {
-			b.Metadata = make(map[string]string, len(kvs))
-		}
-		for k, v := range kvs {
-			b.Metadata[k] = v
-		}
-		c.beads[id] = b
-	}
-	c.mu.Unlock()
-	return nil
-}
-
-// DepAdd adds a dependency and updates the cache.
-func (c *CachingStore) DepAdd(issueID, dependsOnID, depType string) error {
-	if err := c.backing.DepAdd(issueID, dependsOnID, depType); err != nil {
-		return err
-	}
-	c.mu.Lock()
-	deps := c.deps[issueID]
-	for i, d := range deps {
-		if d.DependsOnID == dependsOnID {
-			deps[i].Type = depType
-			c.deps[issueID] = deps
-			c.mu.Unlock()
-			return nil
-		}
-	}
-	c.deps[issueID] = append(deps, Dep{IssueID: issueID, DependsOnID: dependsOnID, Type: depType})
-	c.mu.Unlock()
-	return nil
-}
-
-// DepRemove removes a dependency and updates the cache.
-func (c *CachingStore) DepRemove(issueID, dependsOnID string) error {
-	if err := c.backing.DepRemove(issueID, dependsOnID); err != nil {
-		return err
-	}
-	c.mu.Lock()
-	deps := c.deps[issueID]
-	for i, d := range deps {
-		if d.DependsOnID == dependsOnID {
-			c.deps[issueID] = append(deps[:i], deps[i+1:]...)
-			break
-		}
-	}
-	c.mu.Unlock()
-	return nil
-}
-
-// ── Background reconciler ───────────────────────────────────────────
-
-func (c *CachingStore) reconcileLoop(ctx context.Context) {
-	timer := time.NewTimer(5 * time.Second)
-	defer timer.Stop()
-
-	for {
-		select {
-		case <-ctx.Done():
-			return
-		case <-timer.C:
-		}
-
-		if c.reconciling.CompareAndSwap(false, true) {
-			c.runReconciliation()
-			c.reconciling.Store(false)
-		}
-
-		timer.Reset(c.adaptiveInterval())
-	}
-}
-
-func (c *CachingStore) adaptiveInterval() time.Duration {
-	c.mu.RLock()
-	total := len(c.beads)
-	c.mu.RUnlock()
-	switch {
-	case total >= 5000:
-		return cacheReconcileIntervalLarge
-	case total >= 1000:
-		return cacheReconcileIntervalMedium
-	default:
-		return cacheReconcileIntervalSmall
-	}
-}
-
-func (c *CachingStore) runReconciliation() {
-	start := time.Now()
-	fresh, err := c.backing.List(ListQuery{AllowScan: true})
-	if err != nil {
-		c.mu.Lock()
-		c.syncFailures++
-		if c.syncFailures >= maxCacheSyncFailures && c.state == cacheLive {
-			c.state = cacheDegraded
-			log.Printf("caching-store: degraded after %d consecutive failures", c.syncFailures)
-		}
-		c.mu.Unlock()
-		log.Printf("caching-store: reconcile failed: %v", err)
-		return
-	}
-
-	freshByID := make(map[string]Bead, len(fresh))
-	for _, b := range fresh {
-		freshByID[b.ID] = b
-	}
-
-	c.mu.Lock()
-
-	var adds, removes, updates int64
-	var changedIDs []string
-
-	// Detect new and updated
-	for id, fb := range freshByID {
-		old, exists := c.beads[id]
-		if !exists {
-			c.beads[id] = cloneBead(fb)
-			adds++
-			changedIDs = append(changedIDs, id)
-			c.notifyChangeLocked("bead.created", fb)
-		} else if beadChanged(old, fb) {
-			c.beads[id] = cloneBead(fb)
-			updates++
-			changedIDs = append(changedIDs, id)
-			c.notifyChangeLocked("bead.updated", fb)
-		}
-	}
-
-	// Detect removed
-	for id, old := range c.beads {
-		if _, exists := freshByID[id]; !exists {
-			delete(c.beads, id)
-			delete(c.deps, id)
-			removes++
-			c.notifyChangeLocked("bead.closed", old)
-		}
-	}
-
-	c.syncFailures = 0
-	c.syncedAt = time.Now()
-	if c.state == cacheDegraded {
-		c.state = cacheLive
-		log.Printf("caching-store: recovered to live")
-	}
-
-	durMs := float64(time.Since(start).Microseconds()) / 1000.0
-	c.stats.LastReconcileAt = time.Now()
-	c.stats.LastReconcileMs = durMs
-	c.stats.Adds += adds
-	c.stats.Removes += removes
-	c.stats.Updates += updates
-	c.stats.SyncFailures = c.syncFailures
-	c.updateStatsLocked()
-	c.mu.Unlock()
-
-	// Batch-refresh deps for changed beads (one subprocess call).
-	if len(changedIDs) > 0 {
-		if bdStore, ok := c.backing.(*BdStore); ok {
-			if depMap, err := bdStore.DepListBatch(changedIDs); err == nil {
-				c.mu.Lock()
-				for id, deps := range depMap {
-					c.deps[id] = slices.Clone(deps)
-				}
-				c.mu.Unlock()
-			}
-		} else {
-			// Non-BdStore fallback: per-ID dep fetch.
-			for _, id := range changedIDs {
-				if deps, err := c.backing.DepList(id, "down"); err == nil {
-					c.mu.Lock()
-					c.deps[id] = slices.Clone(deps)
-					c.mu.Unlock()
-				}
-			}
-		}
-	}
-
-	if adds > 0 || removes > 0 || updates > 0 {
-		log.Printf("caching-store: reconciled in %.0fms (+%d -%d ~%d, %d total)",
-			durMs, adds, removes, updates, len(c.beads))
-	}
-}
-
-func (c *CachingStore) notifyChange(eventType string, b Bead) {
-	if c.onChange == nil {
-		return
-	}
-	payload, _ := json.Marshal(b)
-	c.onChange(eventType, b.ID, payload)
-}
-
-func (c *CachingStore) notifyChangeLocked(eventType string, b Bead) {
-	if c.onChange == nil {
-		return
-	}
-	payload, _ := json.Marshal(b)
-	// Unlock before callback to avoid holding the lock during event recording.
-	c.mu.Unlock()
-	c.onChange(eventType, b.ID, payload)
-	c.mu.Lock()
+	c.stats.ProblemCount++
+	c.stats.LastProblemAt = time.Now()
+	c.stats.LastProblem = fmt.Sprintf("%s: %v", op, err)
 }
 
 func (c *CachingStore) updateStatsLocked() {
@@ -839,40 +240,49 @@ func (c *CachingStore) updateStatsLocked() {
 		totalDeps += len(deps)
 	}
 	c.stats.TotalDeps = totalDeps
+	c.stats.SyncFailures = c.syncFailures
 }
 
-func beadChanged(old, fresh Bead) bool {
-	if old.Status != fresh.Status {
-		return true
+func beadIDs(beadMap map[string]Bead) []string {
+	ids := make([]string, 0, len(beadMap))
+	for id := range beadMap {
+		ids = append(ids, id)
 	}
-	if old.Title != fresh.Title {
-		return true
+	return ids
+}
+
+func (c *CachingStore) fetchDepsForIDs(ids []string) (map[string][]Dep, error) {
+	depMap := make(map[string][]Dep)
+	if len(ids) == 0 {
+		return depMap, nil
 	}
-	if old.Assignee != fresh.Assignee {
-		return true
-	}
-	if old.Description != fresh.Description {
-		return true
-	}
-	if len(old.Metadata) != len(fresh.Metadata) {
-		return true
-	}
-	for k, v := range old.Metadata {
-		if fresh.Metadata[k] != v {
-			return true
+
+	if bdStore, ok := c.backing.(*BdStore); ok {
+		batchDeps, err := bdStore.DepListBatch(ids)
+		if err != nil {
+			return depMap, err
 		}
+		for id, deps := range batchDeps {
+			depMap[id] = cloneDeps(deps)
+		}
+		return depMap, nil
 	}
-	return !slices.Equal(old.Labels, fresh.Labels)
+
+	for _, id := range ids {
+		deps, err := c.backing.DepList(id, "down")
+		if err != nil {
+			return depMap, fmt.Errorf("listing deps for %s: %w", id, err)
+		}
+		depMap[id] = cloneDeps(deps)
+	}
+	return depMap, nil
 }
 
-// Delete passes through to the backing store and removes from cache.
-func (c *CachingStore) Delete(id string) error {
-	if err := c.backing.Delete(id); err != nil {
-		return err
+func cloneDeps(deps []Dep) []Dep {
+	if len(deps) == 0 {
+		return nil
 	}
-	c.mu.Lock()
-	delete(c.beads, id)
-	delete(c.deps, id)
-	c.mu.Unlock()
-	return nil
+	cloned := make([]Dep, len(deps))
+	copy(cloned, deps)
+	return cloned
 }
